requirements.txt
deepspeed

diffusers
transformers
peft
datasets
pillow
lpips